+++ app-kotlin/build.gradle
@@ -38,6 +38,9 @@ dependencies {
     def dagger_version = "2.24"
     implementation "com.google.dagger:dagger:$dagger_version"
     kapt "com.google.dagger:dagger-compiler:$dagger_version"
+
+    implementation "io.reactivex.rxjava2:rxjava:2.0.6"
+    implementation "io.reactivex.rxjava2:rxandroid:2.0.1"
 }
 repositories {
     mavenCentral()
