+++ app-kotlin/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         </service>
 
         <activity
-            android:name="com.romellfudi.ussd.sample.MainActivity"
+            android:name="com.romellfudi.ussd.main.MainActivity"
             android:label="@string/app_name"
             android:launchMode="singleTop"
             android:screenOrientation="sensorPortrait"
