+++ app-kotlin/src/main/java/com/romellfudi/ussd/di/component/UIComponent.kt
@@ -8,7 +8,7 @@ package com.romellfudi.ussd.di.component
 
 import com.romellfudi.ussd.di.ActivityQualifier
 import com.romellfudi.ussd.di.module.ActivityModule
-import com.romellfudi.ussd.sample.MainActivity
+import com.romellfudi.ussd.main.MainActivity
 import dagger.Subcomponent
 
 /**
