+++ app-kotlin/src/main/java/com/romellfudi/ussd/main/presenter/MainMVPPresenter.kt
@@ -1,0 +1,7 @@
+package com.romellfudi.ussd.main.presenter
+
+interface MainMVPPresenter {
+    fun call()
+    fun callOverlay()
+    fun callSplashOverlay()
+}
