+++ app-kotlin/src/main/res/layout/activity_main_menu.xml
@@ -12,7 +12,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".sample.MainActivity"
+    tools:context=".main.MainActivity"
     tools:openDrawer="start">
 
     <include
