+++ app/src/main/res/layout/content_main_menu.xml
@@ -15,7 +15,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".sample.MainActivity"
+    tools:context=".main.MainActivity"
     tools:showIn="@layout/app_bar_main_menu">
 
     <FrameLayout
